Microservices/code_runner/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
docker>=5.0.0
pydantic>=1.8.0
python-dotenv>=0.19.0
